[#11030] xCluster,TabletSplitting: Handle master failovers during ProcessSplitTabletResult
Summary:
Changing where xCluster hooks into the tablet splitting code. Will now run ProcessSplitTabletResult after having registered the children, but before actually sending the tablet split rpcs. This ensures that in case of a master failover at any point, the tablet split will be in a pending/started state, and will thus retry via the TabletSplitManager. This removes the need of having loaders to handle any xCluster metadata updates (cdc_state table row entries for children + entries in the consumer-producer tablet mapping).

Additionally, fixing issues with setting the checkpoint for the child tablets. Changing the logic so that we now correctly set the children tablet checkpoints in cdc_state upon reading a SPLIT_OP during a GetChanges for the parent tablet. To avoid race conditions here, we check if the master has created the children entries in the cdc_state table yet, and if not, we only return the records before this opid.

Test Plan:
```
ybd --cxx-test integration-tests_xcluster-tablet-split-itest --gtest_filter XClusterExternalTabletSplitITest.MasterFailoverDuringConsumerPostSplitOps
ybd --cxx-test integration-tests_xcluster-tablet-split-itest --gtest_filter XClusterExternalTabletSplitITest.MasterFailoverDuringProducerPostSplitOps
```

Reviewers: nicolas, asrivastava, rahuldesirazu, hsunder

Reviewed By: asrivastava, rahuldesirazu, hsunder

Subscribers: hsunder, lnguyen, ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D16830

+++ src/yb/integration-tests/cdc_test_util.h
@@ -38,7 +38,7 @@ void VerifyWalRetentionTime(yb::MiniCluster* cluster,
                             uint32_t expected_wal_retention_secs);
 
 Status CorrectlyPollingAllTablets(
-    MiniCluster* cluster, uint32_t num_producer_tablets, MonoDelta timeout);
+    MiniCluster* cluster, size_t num_producer_tablets, MonoDelta timeout);
 } // namespace cdc
 } // namespace yb
